Update API docs to allow changing a subscriber's status
I don't think we should have a "update status" and "unsubscribe_all"
endpoint in our docs since then there are 2 ways to do the exact same
thing.

I first tried doing a `/resubscribe` endpoint but that's a little
confusing: resubscribe to what? To a campaign? All of these docs mention
campaigns all over and subscriptions are very campaign specific. To not
add further confusion, I tried to make the docs clear with a new
endpoint just used to update status.